Code/requirements.txt
numpy==1.24.2
matplotlib==3.8.0
networkx==3.1
PyQt5==5.15.9
tvb-library==2.5